Implements a few issues
