+++ tests/pc-on-line.geo
@@ -1,0 +1,2 @@
+let A, B, C, D = Point();
+ABCD lies_on AB;
